Add support for numpy 2 in semper reader

+++ rsciio/semper/_api.py
@@ -218,7 +218,11 @@ class SemperFormat(object):
         assert label["SEMPER"] == "Semper"
         # Process dimensions:
         for key in ["NCOL", "NROW", "NLAY", "ICCOLN", "ICROWN", "ICLAYN"]:
-            value = 256**2 * label.pop(key + "H") + 256 * label[key][0] + label[key][1]
+            value = (
+                256**2 * np.int32(label.pop(key + "H"))
+                + 256 * label[key][0]
+                + label[key][1]
+            )
             label[key] = value
         # Process date:
         date = "{}-{}-{} {}:{}:{}".format(label["DATE"][0] + 1900, *label["DATE"][1:])
